Change auth type from basic to passport/jwt
UserModel: Extended the userModel with a pre-save hook for encrypting
user password before it is saved. Also, added a helper method which
validates the password provided.
Authentication Middleware: Implemented 3 middlewares which are based on
passport.js, the first helps to register a user and saves the user
information to the DB, it is mounted on the /register route.
The second middleware is mounted on the /login route, and it
is responsible for logging in a user. Successful login returns a JWT
Access Token. Lastly, a third middleware which extracts the
Access Token from subsequent requests, confirms its validity and uses
it and use same to authenticate the /orders
it to authenticate the /orders route.